requirements.txt
pandas
openai==1.1.2
scikit-learn
flask
python-dotenv